SLING-12858 refactoring to reduce duplication
