+++ src/main/java/org/apache/sling/servlethelpers/ByteArrayPart.java
@@ -20,18 +20,13 @@ package org.apache.sling.servlethelpers;
 
 import javax.servlet.http.Part;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Collection;
-
 /**
  * Simple Part implementation backed by an in-memory byte array
  *
  * @deprecated Use {@link JakartaByteArrayPart} instead.
  */
 @Deprecated(since = "2.0.0")
-public class ByteArrayPart implements Part {
+public class ByteArrayPart extends BaseByteArrayPart implements Part {
 
     /**
      * Returns a Builder instance used to create a ByteArrayPart
@@ -42,66 +37,8 @@ public class ByteArrayPart implements Part {
         return new Builder();
     }
 
-    private final byte[] content;
-    private final String name;
-
     private ByteArrayPart(byte[] content, String name) {
-        if (content == null) throw new IllegalArgumentException("content may not be null");
-
-        if (name == null || name.isEmpty()) throw new IllegalArgumentException("name may not be null or empty");
-
-        this.content = content;
-        this.name = name;
-    }
-
-    @Override
-    public InputStream getInputStream() throws IOException {
-        return new ByteArrayInputStream(content);
-    }
-
-    @Override
-    public String getContentType() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String getSubmittedFileName() {
-        return getName();
-    }
-
-    @Override
-    public long getSize() {
-        return content.length;
-    }
-
-    @Override
-    public void write(String fileName) throws IOException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void delete() throws IOException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public String getHeader(String name) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Collection<String> getHeaders(String name) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Collection<String> getHeaderNames() {
-        throw new UnsupportedOperationException();
+        super(content, name);
     }
 
     public static class Builder {
