+++ src/main/java/org/apache/sling/servlethelpers/JakartaByteArrayPart.java
@@ -18,17 +18,12 @@
  */
 package org.apache.sling.servlethelpers;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Collection;
-
 import jakarta.servlet.http.Part;
 
 /**
  * Simple Part implementation backed by an in-memory byte array
  */
-public class JakartaByteArrayPart implements Part {
+public class JakartaByteArrayPart extends BaseByteArrayPart implements Part {
 
     /**
      * Returns a Builder instance used to create a ByteArrayPart
@@ -39,66 +34,8 @@ public class JakartaByteArrayPart implements Part {
         return new Builder();
     }
 
-    private final byte[] content;
-    private final String name;
-
     private JakartaByteArrayPart(byte[] content, String name) {
-        if (content == null) throw new IllegalArgumentException("content may not be null");
-
-        if (name == null || name.isEmpty()) throw new IllegalArgumentException("name may not be null or empty");
-
-        this.content = content;
-        this.name = name;
-    }
-
-    @Override
-    public InputStream getInputStream() throws IOException {
-        return new ByteArrayInputStream(content);
-    }
-
-    @Override
-    public String getContentType() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String getSubmittedFileName() {
-        return getName();
-    }
-
-    @Override
-    public long getSize() {
-        return content.length;
-    }
-
-    @Override
-    public void write(String fileName) throws IOException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void delete() throws IOException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public String getHeader(String name) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Collection<String> getHeaders(String name) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Collection<String> getHeaderNames() {
-        throw new UnsupportedOperationException();
+        super(content, name);
     }
 
     public static class Builder {
