+++ src/main/java/org/apache/sling/servlethelpers/MockHttpSession.java
@@ -21,12 +21,6 @@ package org.apache.sling.servlethelpers;
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpSession;
 
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-import org.apache.commons.collections4.IteratorUtils;
 import org.osgi.annotation.versioning.ConsumerType;
 
 /**
@@ -36,15 +30,9 @@ import org.osgi.annotation.versioning.ConsumerType;
  */
 @ConsumerType
 @Deprecated(since = "2.0.0")
-public class MockHttpSession implements HttpSession {
+public class MockHttpSession extends BaseMockHttpSession implements HttpSession {
 
     private final ServletContext servletContext;
-    private final Map<String, Object> attributeMap = new HashMap<>();
-    private final String sessionID = UUID.randomUUID().toString();
-    private final long creationTime = System.currentTimeMillis();
-    private boolean invalidated = false;
-    private boolean isNew = true;
-    private int maxActiveInterval = 1800;
 
     public MockHttpSession() {
         this.servletContext = newMockServletContext();
@@ -57,108 +45,6 @@ public class MockHttpSession implements HttpSession {
     @Override
     public ServletContext getServletContext() {
         return this.servletContext;
-    }
-
-    @Override
-    public Object getAttribute(final String name) {
-        checkInvalidatedState();
-        return this.attributeMap.get(name);
-    }
-
-    @Override
-    public Enumeration<String> getAttributeNames() {
-        checkInvalidatedState();
-        return IteratorUtils.asEnumeration(this.attributeMap.keySet().iterator());
-    }
-
-    @Override
-    public String getId() {
-        return this.sessionID;
-    }
-
-    @Override
-    public long getCreationTime() {
-        checkInvalidatedState();
-        return this.creationTime;
-    }
-
-    @Override
-    public Object getValue(final String name) {
-        checkInvalidatedState();
-        return getAttribute(name);
-    }
-
-    @Override
-    public String[] getValueNames() {
-        checkInvalidatedState();
-        return this.attributeMap
-                .keySet()
-                .toArray(new String[this.attributeMap.keySet().size()]);
-    }
-
-    @Override
-    public void putValue(final String name, final Object value) {
-        checkInvalidatedState();
-        setAttribute(name, value);
-    }
-
-    @Override
-    public void removeAttribute(final String name) {
-        checkInvalidatedState();
-        this.attributeMap.remove(name);
-    }
-
-    @Override
-    public void removeValue(final String name) {
-        removeAttribute(name);
-    }
-
-    @Override
-    public void setAttribute(final String name, final Object value) {
-        checkInvalidatedState();
-        this.attributeMap.put(name, value);
-    }
-
-    @Override
-    public void invalidate() {
-        checkInvalidatedState();
-        this.invalidated = true;
-    }
-
-    private void checkInvalidatedState() {
-        if (invalidated) {
-            throw new IllegalStateException("Session is already invalidated.");
-        }
-    }
-
-    public boolean isInvalidated() {
-        return invalidated;
-    }
-
-    @Override
-    public boolean isNew() {
-        checkInvalidatedState();
-        return isNew;
-    }
-
-    public void setNew(boolean isNew) {
-        this.isNew = isNew;
-    }
-
-    @Override
-    public long getLastAccessedTime() {
-        checkInvalidatedState();
-        return creationTime;
-    }
-
-    @Override
-    public int getMaxInactiveInterval() {
-        return maxActiveInterval;
-    }
-
-    @Override
-    public void setMaxInactiveInterval(final int interval) {
-        this.maxActiveInterval = interval;
     }
 
     // --- unsupported operations ---
