+++ src/main/java/org/apache/sling/servlethelpers/ResponseBodySupport.java
@@ -21,13 +21,7 @@ package org.apache.sling.servlethelpers;
 import javax.servlet.ServletOutputStream;
 import javax.servlet.WriteListener;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.StandardCharsets;
-import java.util.Objects;
 
 /**
  * Manage response body content.
@@ -35,22 +29,13 @@ import java.util.Objects;
  * @deprecated Use {@link JakartaResponseBodySupport} instead.
  */
 @Deprecated(since = "2.0.0")
-class ResponseBodySupport {
-
-    private ByteArrayOutputStream outputStream;
-    private ServletOutputStream servletOutputStream;
-    private PrintWriter printWriter;
+class ResponseBodySupport extends BaseResponseBodySupport<ServletOutputStream> {
 
     public ResponseBodySupport() {
         reset();
     }
 
-    public void reset() {
-        outputStream = new ByteArrayOutputStream();
-        servletOutputStream = null;
-        printWriter = null;
-    }
-
+    @Override
     public ServletOutputStream getOutputStream() {
         if (servletOutputStream == null) {
             servletOutputStream = new ServletOutputStream() {
@@ -71,42 +56,5 @@ class ResponseBodySupport {
             };
         }
         return servletOutputStream;
-    }
-
-    public PrintWriter getWriter(String charset) {
-        if (printWriter == null) {
-            try {
-                printWriter = new PrintWriter(new OutputStreamWriter(getOutputStream(), defaultCharset(charset)));
-            } catch (UnsupportedEncodingException ex) {
-                throw new RuntimeException("Unsupported encoding: " + defaultCharset(charset), ex);
-            }
-        }
-        return printWriter;
-    }
-
-    public byte[] getOutput() {
-        if (printWriter != null) {
-            printWriter.flush();
-        }
-        if (servletOutputStream != null) {
-            try {
-                servletOutputStream.flush();
-            } catch (IOException ex) {
-                // ignore
-            }
-        }
-        return outputStream.toByteArray();
-    }
-
-    public String getOutputAsString(String charset) {
-        try {
-            return new String(getOutput(), defaultCharset(charset));
-        } catch (UnsupportedEncodingException ex) {
-            throw new RuntimeException("Unsupported encoding: " + defaultCharset(charset), ex);
-        }
-    }
-
-    private String defaultCharset(String charset) {
-        return Objects.toString(charset, StandardCharsets.UTF_8.name());
     }
 }
