+++ src/main/java/org/apache/sling/servlethelpers/internalrequests/InternalRequest.java
@@ -24,11 +24,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.Reader;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.OptionalInt;
-import java.util.stream.Collectors;
 
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
@@ -38,8 +34,6 @@ import org.apache.sling.servlethelpers.MockRequestPathInfo;
 import org.apache.sling.servlethelpers.MockSlingHttpServletRequest;
 import org.apache.sling.servlethelpers.MockSlingHttpServletResponse;
 import org.jetbrains.annotations.NotNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
 
 /** Fluent helper for Sling internal requests.
@@ -53,104 +47,48 @@ import org.slf4j.MDC;
  * @deprecated Use {@link JakartaInternalRequest} instead.
  */
 @Deprecated(since = "2.0.0")
-public abstract class InternalRequest {
-    protected final ResourceResolver resourceResolver;
-    protected final String path;
-    protected String selectorString;
-    protected String extension;
-    protected String requestMethod = DEFAULT_METHOD;
-    protected String contentType;
-    private Reader bodyReader;
-    private boolean explicitStatusCheck;
-    private Map<String, Object> parameters = new HashMap<>();
+public abstract class InternalRequest extends BaseInternalRequest {
     private MockSlingHttpServletRequest request;
     private MockSlingHttpServletResponse response;
 
-    protected final Logger log = LoggerFactory.getLogger(getClass());
-
-    public static final String DEFAULT_METHOD = "GET";
-
-    /** An slf4j MDC value is set at this key with request information.
-     *  That's useful for troubleshooting when using multiple internal
-     *  requests in the context of a single HTTP request.
-     */
-    public static final String MDC_KEY = "sling." + InternalRequest.class.getSimpleName();
-
     /** Clients use subclasses of this one  */
     protected InternalRequest(@NotNull ResourceResolver resourceResolver, @NotNull String path) {
-        checkNotNull(ResourceResolver.class, resourceResolver);
-        checkNotNull("path", path);
-        this.resourceResolver = resourceResolver;
-        this.path = path;
+        super(resourceResolver, path);
     }
 
-    protected void checkNotNull(String info, Object candidate) {
-        if (candidate == null) {
-            throw new IllegalArgumentException(info + " is null");
-        }
-    }
-
-    protected void checkNotNull(Class<?> clazz, Object candidate) {
-        checkNotNull(clazz.getSimpleName(), candidate);
-    }
-
-    /** Set the HTTP request method to use - defaults to GET */
+    @Override
     public InternalRequest withRequestMethod(String method) {
-        this.requestMethod = method.toUpperCase();
-        return this;
+        return (InternalRequest) super.withRequestMethod(method);
     }
 
-    /** Set the HTTP request's Content-Type */
+    @Override
     public InternalRequest withContentType(String contentType) {
-        this.contentType = contentType;
-        return this;
+        return (InternalRequest) super.withContentType(contentType);
     }
 
-    /** Use the supplied Reader as the request's body content */
+    @Override
     public InternalRequest withBody(Reader bodyContent) {
-        bodyReader = bodyContent;
-        return this;
+        return (InternalRequest) super.withBody(bodyContent);
     }
 
-    /** Sets the optional selectors of the internal request, which influence
-     *  the Servlet/Script resolution.
-     */
+    @Override
     public InternalRequest withSelectors(String... selectors) {
-        if (selectors == null) {
-            return this;
-        }
-        StringBuilder sb = new StringBuilder();
-        Arrays.stream(selectors)
-                .forEach(sel -> sb.append(sb.length() == 0 ? "" : ".").append(sel));
-        selectorString = sb.toString();
-        return this;
+        return (InternalRequest) super.withSelectors(selectors);
     }
 
-    /** Sets the optional extension of the internal request, which influence
-     *  the Servlet/Script resolution.
-     */
+    @Override
     public InternalRequest withExtension(String extension) {
-        this.extension = extension;
-        return this;
+        return (InternalRequest) super.withExtension(extension);
     }
 
-    /** Set a request parameter */
+    @Override
     public InternalRequest withParameter(String key, Object value) {
-        if (key != null && value != null) {
-            parameters.put(key, value);
-        } else {
-            throw new IllegalArgumentException("Null key or value");
-        }
-        return this;
+        return (InternalRequest) super.withParameter(key, value);
     }
 
-    /** Add the supplied request parameters to the current ones */
+    @Override
     public InternalRequest withParameters(Map<String, Object> additionalParameters) {
-        if (additionalParameters != null) {
-            parameters.putAll(additionalParameters);
-        }
-
-        return this;
+        return (InternalRequest) super.withParameters(additionalParameters);
     }
 
     /** Execute the internal request. Can be called right after
@@ -159,6 +97,7 @@ public abstract class InternalRequest {
      *  @throws IOException if the request was already executed,
      *      or if an error occurs during execution.
      */
+    @Override
     public final InternalRequest execute() throws IOException {
         if (request != null) {
             throw new IOException("Request was already executed");
@@ -199,9 +138,6 @@ public abstract class InternalRequest {
         return this;
     }
 
-    /** Provide the Resource to use to execute the request */
-    protected abstract Resource getExecutionResource();
-
     /** Execute the supplied Request */
     protected abstract void delegateExecute(
             SlingHttpServletRequest request, SlingHttpServletResponse response, ResourceResolver resourceResolver)
@@ -213,37 +149,9 @@ public abstract class InternalRequest {
         }
     }
 
-    /** After executing the request, checks that the request status is one
-     *  of the supplied values.
-     *
-     *  If this is not called before methods that access the response, a check
-     *  for a 200 OK status is done automatically unless this was called
-     *  with no arguments before.
-     *
-     *  This makes sure a status check is done or explicitly disabled.
-     *
-     *  @param acceptableValues providing no values means "don't care"
-     *  @throws IOException if status doesn't match any of these values
-     */
+    @Override
     public InternalRequest checkStatus(int... acceptableValues) throws IOException {
-        assertRequestExecuted();
-        explicitStatusCheck = true;
-
-        if (acceptableValues == null || acceptableValues.length == 0) {
-            return this;
-        }
-
-        final int actualStatus = getStatus();
-        final OptionalInt found = Arrays.stream(acceptableValues)
-                .filter(expected -> expected == actualStatus)
-                .findFirst();
-        if (!found.isPresent()) {
-            String sb = Arrays.stream(acceptableValues)
-                    .mapToObj(String::valueOf) // Convert each int to its String representation
-                    .collect(Collectors.joining(",")); // Join the strings with a comma delimiter
-            throw new IOException("Unexpected response status " + actualStatus + ", expected one of '" + sb + "'");
-        }
-        return this;
+        return (InternalRequest) super.checkStatus(acceptableValues);
     }
 
     /** If response status hasn't been explicitly checked, ensure it's 200 */
@@ -262,6 +170,7 @@ public abstract class InternalRequest {
      *
      *  @throws IOException if the actual content-type doesn't match the expected one
      */
+    @Override
     public InternalRequest checkResponseContentType(String contentType) throws IOException {
         assertRequestExecuted();
         if (!contentType.equals(response.getContentType())) {
@@ -273,6 +182,7 @@ public abstract class InternalRequest {
     /** Return the response status. The execute method must be called before this one.
      *  @throws IOException if the request hasn't been executed yet
      */
+    @Override
     public int getStatus() throws IOException {
         assertRequestExecuted();
         return response.getStatus();
