+++ src/main/java/org/apache/sling/servlethelpers/internalrequests/JakartaInternalRequest.java
@@ -21,11 +21,7 @@ package org.apache.sling.servlethelpers.internalrequests;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.Reader;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.OptionalInt;
-import java.util.stream.Collectors;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletResponse;
@@ -37,8 +33,6 @@ import org.apache.sling.servlethelpers.MockJakartaSlingHttpServletRequest;
 import org.apache.sling.servlethelpers.MockJakartaSlingHttpServletResponse;
 import org.apache.sling.servlethelpers.MockRequestPathInfo;
 import org.jetbrains.annotations.NotNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
 
 /** Fluent helper for Sling internal requests.
@@ -49,103 +43,48 @@ import org.slf4j.MDC;
  *  executes an HTTP request and another one that's faster by
  *  calling Servlets or Scripts directly.
  */
-public abstract class JakartaInternalRequest {
-    protected final ResourceResolver resourceResolver;
-    protected final String path;
-    protected String selectorString;
-    protected String extension;
-    protected String requestMethod = DEFAULT_METHOD;
-    protected String contentType;
-    private Reader bodyReader;
-    private boolean explicitStatusCheck;
-    private Map<String, Object> parameters = new HashMap<>();
+public abstract class JakartaInternalRequest extends BaseInternalRequest {
     private MockJakartaSlingHttpServletRequest request;
     private MockJakartaSlingHttpServletResponse response;
 
-    protected final Logger log = LoggerFactory.getLogger(getClass());
-
-    public static final String DEFAULT_METHOD = "GET";
-
-    /** An slf4j MDC value is set at this key with request information.
-     *  That's useful for troubleshooting when using multiple internal
-     *  requests in the context of a single HTTP request.
-     */
-    public static final String MDC_KEY = "sling." + JakartaInternalRequest.class.getSimpleName();
-
     /** Clients use subclasses of this one  */
     protected JakartaInternalRequest(@NotNull ResourceResolver resourceResolver, @NotNull String path) {
-        checkNotNull(ResourceResolver.class, resourceResolver);
-        checkNotNull("path", path);
-        this.resourceResolver = resourceResolver;
-        this.path = path;
+        super(resourceResolver, path);
     }
 
-    protected void checkNotNull(String info, Object candidate) {
-        if (candidate == null) {
-            throw new IllegalArgumentException(info + " is null");
-        }
-    }
-
-    protected void checkNotNull(Class<?> clazz, Object candidate) {
-        checkNotNull(clazz.getSimpleName(), candidate);
-    }
-
-    /** Set the HTTP request method to use - defaults to GET */
+    @Override
     public JakartaInternalRequest withRequestMethod(String method) {
-        this.requestMethod = method.toUpperCase();
-        return this;
+        return (JakartaInternalRequest) super.withRequestMethod(method);
     }
 
-    /** Set the HTTP request's Content-Type */
+    @Override
     public JakartaInternalRequest withContentType(String contentType) {
-        this.contentType = contentType;
-        return this;
+        return (JakartaInternalRequest) super.withContentType(contentType);
     }
 
-    /** Use the supplied Reader as the request's body content */
+    @Override
     public JakartaInternalRequest withBody(Reader bodyContent) {
-        bodyReader = bodyContent;
-        return this;
+        return (JakartaInternalRequest) super.withBody(bodyContent);
     }
 
-    /** Sets the optional selectors of the internal request, which influence
-     *  the Servlet/Script resolution.
-     */
+    @Override
     public JakartaInternalRequest withSelectors(String... selectors) {
-        if (selectors == null) {
-            return this;
-        }
-        StringBuilder sb = new StringBuilder();
-        Arrays.stream(selectors)
-                .forEach(sel -> sb.append(sb.length() == 0 ? "" : ".").append(sel));
-        selectorString = sb.toString();
-        return this;
+        return (JakartaInternalRequest) super.withSelectors(selectors);
     }
 
-    /** Sets the optional extension of the internal request, which influence
-     *  the Servlet/Script resolution.
-     */
+    @Override
     public JakartaInternalRequest withExtension(String extension) {
-        this.extension = extension;
-        return this;
+        return (JakartaInternalRequest) super.withExtension(extension);
     }
 
-    /** Set a request parameter */
+    @Override
     public JakartaInternalRequest withParameter(String key, Object value) {
-        if (key != null && value != null) {
-            parameters.put(key, value);
-        } else {
-            throw new IllegalArgumentException("Null key or value");
-        }
-        return this;
+        return (JakartaInternalRequest) super.withParameter(key, value);
     }
 
-    /** Add the supplied request parameters to the current ones */
+    @Override
     public JakartaInternalRequest withParameters(Map<String, Object> additionalParameters) {
-        if (additionalParameters != null) {
-            parameters.putAll(additionalParameters);
-        }
-        return this;
+        return (JakartaInternalRequest) super.withParameters(additionalParameters);
     }
 
     /** Execute the internal request. Can be called right after
@@ -194,9 +133,6 @@ public abstract class JakartaInternalRequest {
         return this;
     }
 
-    /** Provide the Resource to use to execute the request */
-    protected abstract Resource getExecutionResource();
-
     /** Execute the supplied Request */
     protected abstract void delegateExecute(
             SlingJakartaHttpServletRequest request,
@@ -210,37 +146,9 @@ public abstract class JakartaInternalRequest {
         }
     }
 
-    /** After executing the request, checks that the request status is one
-     *  of the supplied values.
-     *
-     *  If this is not called before methods that access the response, a check
-     *  for a 200 OK status is done automatically unless this was called
-     *  with no arguments before.
-     *
-     *  This makes sure a status check is done or explicitly disabled.
-     *
-     *  @param acceptableValues providing no values means "don't care"
-     *  @throws IOException if status doesn't match any of these values
-     */
+    @Override
     public JakartaInternalRequest checkStatus(int... acceptableValues) throws IOException {
-        assertRequestExecuted();
-        explicitStatusCheck = true;
-
-        if (acceptableValues == null || acceptableValues.length == 0) {
-            return this;
-        }
-
-        final int actualStatus = getStatus();
-        final OptionalInt found = Arrays.stream(acceptableValues)
-                .filter(expected -> expected == actualStatus)
-                .findFirst();
-        if (!found.isPresent()) {
-            String sb = Arrays.stream(acceptableValues)
-                    .mapToObj(String::valueOf) // Convert each int to its String representation
-                    .collect(Collectors.joining(",")); // Join the strings with a comma delimiter
-            throw new IOException("Unexpected response status " + actualStatus + ", expected one of '" + sb + "'");
-        }
-        return this;
+        return (JakartaInternalRequest) super.checkStatus(acceptableValues);
     }
 
     /** If response status hasn't been explicitly checked, ensure it's 200 */
