+++ src/main/java/org/apache/sling/servlethelpers/internalrequests/package-info.java
@@ -19,5 +19,5 @@
 /**
  * Helpers for internal Sling requests
  */
-@org.osgi.annotation.versioning.Version("1.1.0")
+@org.osgi.annotation.versioning.Version("2.0.0")
 package org.apache.sling.servlethelpers.internalrequests;
